fix: make the updater compile on all platforms
