+++ internal/updater/updater_stubs.go
@@ -1,0 +1,7 @@
+//go:build !windows
+
+package updater
+
+func launchElevated(exePath string, args string, workingDir string) error {
+	return nil
+}
